# resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Omicflows",
    "version": "1.0-SNAPSHOT",
    "title": "Omicflows backend",
    "contact": {
      "name": "rsperoni@mgcoders.com"
    }
  },
  "host": "omicflows-back.mgcoders.com",
  "basePath": "/api",
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List jobs",
        "description": "",
        "operationId": "get",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Job"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "New job",
        "description": "",
        "operationId": "newJob",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Workflow"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Job"
            }
          }
        }
      }
    },
    "/jobs/{jobId}": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Add simple resource to Job",
        "description": "",
        "operationId": "addSimpleResourceToJob",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/JobResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Job"
            }
          }
        }
      },
      "put": {
        "tags": [
          "jobs"
        ],
        "summary": "Close Job",
        "description": "",
        "operationId": "closeJob",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Job"
            }
          }
        }
      }
    },
    "/jobs/{workflowId}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get jobs by workflow",
        "description": "",
        "operationId": "getByWorkflow",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "workflowId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Job"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "WorkflowIn": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "sourceMappedToolName": {
          "type": "string"
        },
        "sourceMappedPortName": {
          "type": "string"
        },
        "mapped": {
          "type": "boolean",
          "default": false
        },
        "schema": {
          "type": "string"
        }
      }
    },
...
    "WorkflowStep": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "json": {
          "type": "string"
        },
        "neededInputs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowIn"
          }
        },
        "neededOutputs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowOut"
          }
        }
      }
    },
    "JobResource": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "get_class": {
          "type": "string"
        }
      }
    },
    "Workflow": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "complete": {
          "type": "boolean",
          "default": false
        },
        "steps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowStep"
          }
        },
        "neededInputs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowIn"
          }
        },
        "neededOutputs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WorkflowOut"
          }
        }
      }
    },
...
    "Job": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "workflowId": {
          "type": "string"
        },
        "workflow": {
          "$ref": "#/definitions/Workflow"
        },
        "resources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JobResource"
          }
        },
        "ready": {
          "type": "boolean",
          "default": false
        },
        "inputsFilePath": {
          "type": "string"
        },
        "jsonWorkflowFilePath": {
          "type": "string"
        },
        "yamlWorkflowFilePath": {
          "type": "string"
        }
      }
    },
    "WorkflowOut": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "schema": {
          "type": "string"
        }
      }
    },
...
  }
}